better syntax

+++ source/common/common/thread.cc
@@ -32,7 +32,7 @@ int32_t Thread::currentThreadId() {
 #ifdef __linux__
   return syscall(SYS_gettid);
 #elif defined(__FreeBSD__)
-  return pthread_getthreadid_np();
+  return static_cast<int32_t>(pthread_getthreadid_np());
 #elif defined(__APPLE__)
   uint64_t tid;
   pthread_threadid_np(NULL, &tid);
